fix deletefiles spamming console.
removed invalid comment "must be backslash for deleteFiles" as it only applies to windows in deleteFiles.

Removed relative paths and replaced them with absolute paths in deleteFiles and also wherever it is used (it didnt work if i tried to get the absolute path in the function for some reason).

Added check for accidental deletion of root in deleteFiles.

added error stream printing to deleteFiles